requirements.txt
streamlit==1.32.0
pandas==2.2.1
google-genai==0.3.0
